Exit with non-zero code if docs plugin has error
Prior to this commit running a documentation plugin resulted in a zero
exit code even if the plugin explicitly returned an error. This meant
that in a CI/CD pipeline, for instance, the pipeline couldn't check the
exit code and instead had to check the console or log output, which was
brittle.

Fixes #2080

Signed-off-by: John Boyes <154404+johnboyes@users.noreply.github.com>